mount.nfs: Eliminate compiler warnings in utils/mount/version.h
Clean up.

In file included from mount.c:50:
version.h: In function linux_version_code:
version.h:48: warning: conversion to unsigned int from int may
    change the sign of the result
version.h:48: warning: conversion to unsigned int from int may
    change the sign of the result
version.h:48: warning: conversion to unsigned int from int may
    change the sign of the result

+++ utils/mount/version.h
@@ -42,9 +42,9 @@ static inline unsigned int linux_version_code(void)
 	if (uname(&my_utsname))
 		return 0;
 
-	p = atoi(strtok(my_utsname.release, "."));
-	q = atoi(strtok(NULL, "."));
-	r = atoi(strtok(NULL, "."));
+	p = (unsigned int)atoi(strtok(my_utsname.release, "."));
+	q = (unsigned int)atoi(strtok(NULL, "."));
+	r = (unsigned int)atoi(strtok(NULL, "."));
 	return MAKE_VERSION(p, q, r);
 }
